requirements.txt
streamlit==1.28.1
spacy==3.7.2
transformers==4.35.0
torch==2.2.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
pdfminer.six==20221105
PyMuPDF==1.23.8
plotly==5.17.0
openai==1.3.5
python-dotenv==1.0.0
nltk==3.8.1
